Handle any size numbers in DWORD and QWORD values. This may be technically incorrect, but it eleminates issues with converting between signed and unsigned ints and longs, at the extreme upper range.

+++ src/org/joval/intf/windows/registry/IDwordValue.java
@@ -3,6 +3,8 @@
 
 package org.joval.intf.windows.registry;
 
+import java.math.BigInteger;
+
 /**
  * Interface to a Windows registry DWORD value.
  *
@@ -13,5 +15,5 @@ public interface IDwordValue extends IValue {
     /**
      * Get the data.
      */
-    public int getData();
+    public BigInteger getData();
 }
